Added static typing, moved source files to project folder, separated tests into two files
