+++ .github/workflows/check_types.yml
@@ -1,0 +1,19 @@
+name: Check types
+
+on:
+  [ push, pull_request ]
+
+jobs:
+  type-check:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install mypy
+          pip install -r requirements.txt
+
+      - name: Run mypy
+        run: |
+          mypy .
