+++ .github/workflows/tests.yml
@@ -1,7 +1,6 @@
-name: Check code style
+name: Run tests
 
 on:
-  # Triggers the workflow on push or pull request events
   [ push, pull_request ]
 
 jobs:
@@ -9,19 +8,6 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.x' 
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-
       - name: Run tests
         run: |
           pytest
